requirements_dev.txt
pip==19.0.3
bumpversion==0.5.3
wheel==0.33.1
watchdog==0.9.0
flake8==3.7.7
tox==3.7.0
coverage==4.5.2
Sphinx==1.8.4
twine==1.13.0

pytest==4.3.0
pytest-runner==4.4
pytest-mock==1.10.1
hypothesis==4.8.0
requests-toolbelt==0.9.1
